added: Feed, Stories, Story, InputBox, firebase.js
Feed component has Stories
InputBox is for user to post text and/or an image
Sonny's choice of SDK from firebase was a combination of Node.js (client) and browser js ver. 9 for Storage, but there are some issues with autocompletion with Node.js in later verions of the SDK (firebase 9.6).
For consistency, I have chosen browser js ver. 9 for both Firestore and Storage, and also for initialization of the firebase app.